better naming (system_ffi etc)
